Added an example around VRAM usage
This shall simplify understanding of the two data registers and how to work with them. I found out during my examples, that the ADRSEL have to be set first before setting the address. That might be a symptom of the Emulator, but I created weird effect if Setting ADRSEL last.  Added a small piece of example code for better understanding.